# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shivafer Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channel/posts/{id}/view:
    post:
      operationId: incrementChannelPostView
      tags: [channel]
      summary: Increment view count for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Reels (public) ─────────────────────────────────────
...
  /admin/channel/posts:
    get:
      operationId: adminGetChannelPosts
      tags: [admin]
      summary: Admin - list all channel posts
      security:
        - adminAuth: []
      responses:
        "200":
          description: All channel posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelPost"
    post:
      operationId: adminCreateChannelPost
      tags: [admin]
      summary: Admin - create channel post
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPost"
  /admin/channel/posts/{id}:
    put:
      operationId: adminUpdateChannelPost
      tags: [admin]
      summary: Admin - update channel post
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelPostInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPost"
    delete:
      operationId: adminDeleteChannelPost
      tags: [admin]
      summary: Admin - delete channel post
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/channel/posts/{id}/pin:
    patch:
      operationId: adminPinChannelPost
      tags: [admin]
      summary: Admin - pin/unpin channel post
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelPostPinInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPost"

  # ─── Admin Reels ─────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ChannelPost:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        mediaUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          nullable: true
        isPinned:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - isPinned
        - viewCount
        - createdAt
    ChannelPostInput:
      type: object
      properties:
        content:
          type: string
        mediaUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          nullable: true
        isPinned:
          type: boolean
      required:
        - content
...
    ChannelPostPinInput:
      type: object
      properties:
        isPinned:
          type: boolean
      required:
        - isPinned